+++ R/data_overview_module.R
@@ -2,34 +2,28 @@
 # ui --------------------------------------------------------------------------
 data_overview_ui <- function(id) {
   ns <- NS(id)
-  tagList(
-    card_body(gt_output(ns('gt')),
-              style = "margin-top: -24px !important;
-                       margin-bottom: -24px !important;
-                       padding-top: -24px !important;
-                       padding-bottom: -24px !important;
-              "),
-    fluidRow(
-      column(2, numericInput(ns('size_sample'), 'Number of rows', 100, 100, 1e4, 100)),
-      column(2, radioGroupButtons(
-        ns('radio_sample'), 'Show',
-        c('First rows' = 'first', 'Sample' = 'sample'),
-        size = 'sm', individual = T)),
-      style = "margin-top: -16px !important; margin-bottom: -16px !important;"
-    ),
+  card(
+    card_body(gt_output(ns('gt'))),
+    card_footer(
+      fluidRow(
+        column(2, numericInput(ns('size_sample'), 'Number of rows', 500, 0, 1e4, 500)),
+        column(2, radioGroupButtons(
+          ns('radio_sample'), 'Show',
+          c('First rows' = 'first', 'Sample' = 'sample'),
+          size = 'sm', individual = T))
+      )
+    )
   )
 }
 
 # server ----------------------------------------------------------------------
-data_overview_server <- function(id, df, triggers) {
+data_overview_server <- function(id, df) {
   moduleServer(id, function(input, output, session) {
 
     output$gt <- render_gt({
         req(input$size_sample)
 
         validate(need(input$size_sample > 0, 'Number of rows must be > 0'))
-
-        triggers()
 
         n_show <- max(1, input$size_sample)
         n_show <- min(n_show, nrow(df()))