+++ R/rename_cols_module.R
@@ -11,7 +11,6 @@ rename_cols_ui <- function(id) {
     ),
     card_footer(btn_task(ns('btn_rename'), 'Rename Variable', icon('check')))
   )
-
 }
 
 # server ----------------------------------------------------------------------
@@ -19,19 +18,15 @@ rename_cols_server <- function(id, input_df) {
   moduleServer(id, function(input, output, session) {
     ns <- NS(id)
 
+    # Reactive to get column names
+    df_names <- reactive(input_df() |> names())
+
     df <- reactiveValues()
     observe({
-      req(input_df())
       df$df_active <- input_df()
     })
 
-    df_names <- reactive({
-      req(df$df_active)
-      df$df_active |> names()
-    })
-
     output$ui_var_sel <- renderUI({
-      req(df$df_active)
       selectInput(ns('vars_sel'), 'Variable', c('', df_names()))
     })
 
@@ -42,13 +37,15 @@ rename_cols_server <- function(id, input_df) {
       } else {
         if(is_valid_name(input$txt_new_name) &&
            input$txt_new_name %notin% df_names()){
-          names(df$df_active)[names(df$df_active) == input$vars_sel] <- input$txt_new_name
+
+          temp <- copy(df$df_active)
+
+          setnames(temp, input$vars_sel, input$txt_new_name)
+
+          df$df_active <- copy(temp)
+		  rm(temp)
+
           msg('Rename columns: OK')
-
-          output$ui_var_sel <- renderUI({
-            selectInput(ns('vars_sel'), 'Variable', c('', df_names()))
-          })
-
         } else {
           msg_error('New name is not valid or already in use')
         }